Refactored EventForm
